endpoint overall ranking

+++ application-impl/src/main/java/ch/heigvd/amt/projet2/entities/ResultOverAll.java
@@ -4,6 +4,6 @@ import java.util.UUID;
 
 public interface ResultOverAll {
     UUID getIdUser();
-    Long getTotalBadge();
-    Long getTotalPointScale();
+    int getTotalBadge();
+    int getTotalPointScale();
 }
